player_info no longer resets to $00 at beginning of game loop.
Attempts were made to reset specific bits when corresponding actions are not supposed to be in use. Not completely successful. Player now moves upwards when left or right are held. Goal is to handle actions in a way that is not directly tied to button presses (Player should not only move up when A button is held. This will also likely help when implementing the latter half of a jump arc too, so player can fall down instead of rising forever).